+++ inputs/ValidateUserInput.go
@@ -1,7 +1,7 @@
 package inputs
 
 //validate user inputs : num1 and num2
-func ValidateUserInput(num1,num2 int)bool{
+func ValidateUserInput(num1,num2 float64)bool{
     return  num1 > 0 && num2 > 0
     //Returns true if the inputs are valid
     //Returns false if the inputs are invalid
