+++ mathutils/Sub.go
@@ -1,6 +1,6 @@
 package mathutils
 
 
-func Subtract(num1,num2 int) (int){
+func Subtract(num1,num2 float64) (float64){
     return num1 - num2
 }
